+++ libcnb-test/src/pack.rs
@@ -1,7 +1,6 @@
-use crate::PackResult;
 use std::collections::BTreeMap;
 use std::path::PathBuf;
-use std::process::{Command, Output};
+use std::process::Command;
 use tempfile::TempDir;
 
 /// Represents a `pack build` command.
@@ -161,43 +160,6 @@ impl From<PackSbomDownloadCommand> for Command {
 
         command
     }
-}
-
-/// Runs the given pack command, panicking with user-friendly error messages when errors occur.
-pub(crate) fn run_pack_command<C: Into<Command>>(
-    command: C,
-    expected_result: &PackResult,
-) -> Output {
-    let output = command.into()
-        .output()
-        .unwrap_or_else(|io_error| {
-            if io_error.kind() == std::io::ErrorKind::NotFound {
-                panic!("External `pack` command not found. Install Pack CLI and ensure it is on PATH: https://buildpacks.io/docs/install-pack");
-            } else {
-                panic!("Could not spawn external `pack` process: {io_error}");
-            };
-        });
-
-    if (expected_result == &PackResult::Success && !output.status.success())
-        || (expected_result == &PackResult::Failure && output.status.success())
-    {
-        panic!(
-            "pack command unexpectedly {} with exit-code {}!\n\npack stdout:\n{}\n\npack stderr:\n{}",
-            if output.status.success() {
-                "succeeded"
-            } else {
-                "failed"
-            },
-            output
-                .status
-                .code()
-                .map_or(String::from("<unknown>"), |exit_code| exit_code.to_string()),
-            String::from_utf8_lossy(&output.stdout),
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    output
 }
 
 #[cfg(test)]